Promote `Logger` to `main` 🚧
